# usuariosflask/openapi_server/openapi/openapi.yaml
openapi: 3.0.3
info:
  description: API de Netflux realizada por el grupo 3 de la asignatura ASEE. Microservicio
    de **Usuarios**.
  license:
    name: Copyleft
    url: https://url/a/la/licencia
  title: Netflux Grupo 3 - OpenAPI 3.0
  version: 0.0.1
servers:
- url: https://path/api/v1
...
paths:
...
  /usuarios:
    get:
      description: Obtiene una lista de todos los usuarios disponibles en el sistema.
      operationId: listar_usuarios
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Usuario'
                type: array
          description: Lista de usuarios obtenida exitosamente
        "500":
          description: Error interno del servidor
      summary: Listar todos los usuarios
      tags:
      - Usuario
      x-openapi-router-controller: openapi_server.controllers.usuario_controller
    post:
      description: Crea un nuevo usuario con la información proporcionada.
      operationId: crear_usuario
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Usuario'
        description: Objeto del usuario a crear
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Usuario'
          description: Usuario creado exitosamente
        "400":
          description: Solicitud inválida
        "500":
          description: Error interno del servidor
      summary: Crear un nuevo usuario
      tags:
      - Usuario
      x-openapi-router-controller: openapi_server.controllers.usuario_controller
  /usuarios/{userId}:
    delete:
      description: Elimina un usuario específico por su ID.
      operationId: eliminar_usuario
      parameters:
      - description: ID del usuario a eliminar
        explode: false
        in: path
        name: userId
        required: true
        schema:
          example: 1
          format: int64
          type: integer
        style: simple
      responses:
        "204":
          description: Usuario eliminado exitosamente
        "404":
          description: Usuario no encontrado
        "500":
          description: Error interno del servidor
      summary: Eliminar un usuario
      tags:
      - Usuario
      x-openapi-router-controller: openapi_server.controllers.usuario_controller
    put:
      description: Actualiza la información de un usuario específico por su ID.
      operationId: actualizar_usuario
      parameters:
      - description: ID del usuario a actualizar
        explode: false
        in: path
        name: userId
        required: true
        schema:
          example: 1
          format: int64
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UsuarioUpdate'
        description: Objeto del usuario con la información actualizada
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Usuario'
          description: Usuario actualizado exitosamente
        "400":
          description: Solicitud inválida
        "404":
          description: Usuario no encontrado
        "500":
          description: Error interno del servidor
      summary: Actualizar un usuario existente
      tags:
      - Usuario
      x-openapi-router-controller: openapi_server.controllers.usuario_controller
  /usuarios/{userId}/password:
    patch:
      description: Actualiza la contraseña de un usuario específico por su ID.
      operationId: actualizar_password
      parameters:
      - description: ID del usuario a actualizar
        explode: false
        in: path
        name: userId
        required: true
        schema:
          example: 1
          format: int64
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              type: string
        description: Nueva contraseña del usuario
        required: true
      responses:
        "200":
          description: Contraseña actualizada exitosamente
        "400":
          description: Solicitud inválida
        "404":
          description: Usuario no encontrado
        "500":
          description: Error interno del servidor
      summary: Actualizar la contraseña de un usuario
      tags:
      - Usuario
      x-openapi-router-controller: openapi_server.controllers.usuario_controller
...
components:
  schemas:
    Usuario:
      example:
        dispositivos:
        - Móvil
        - TV
        correo_electronico: juan@gmail.com
        plan_suscripcion: Básico
        perfiles:
        - user_id: 3
          preferencias_contenido:
            languages:
            - Espanol
            genres:
            - Action
            - Supernatural
          foto_perfil: path/to/image
          id: 1
          nombre: Euler
          historial_vistos:
          - numero: 1
            titulo: Piloto
            duracion: 58
            id: 1
            sinopsis: Walter White toma una decisión radical tras ser diagnosticado
              con cáncer de pulmón.
          - numero: 1
            titulo: Piloto
            duracion: 58
            id: 1
            sinopsis: Walter White toma una decisión radical tras ser diagnosticado
              con cáncer de pulmón.
          mi_lista:
          - numero: 1
            titulo: Piloto
            duracion: 58
            id: 1
            sinopsis: Walter White toma una decisión radical tras ser diagnosticado
              con cáncer de pulmón.
          - numero: 1
            titulo: Piloto
            duracion: 58
            id: 1
            sinopsis: Walter White toma una decisión radical tras ser diagnosticado
              con cáncer de pulmón.
        - user_id: 3
          preferencias_contenido:
            languages:
            - Espanol
            genres:
            - Action
            - Supernatural
          foto_perfil: path/to/image
          id: 1
          nombre: Euler
          historial_vistos:
          - numero: 1
            titulo: Piloto
            duracion: 58
            id: 1
            sinopsis: Walter White toma una decisión radical tras ser diagnosticado
              con cáncer de pulmón.
          - numero: 1
            titulo: Piloto
            duracion: 58
            id: 1
            sinopsis: Walter White toma una decisión radical tras ser diagnosticado
              con cáncer de pulmón.
          mi_lista:
          - numero: 1
            titulo: Piloto
            duracion: 58
            id: 1
            sinopsis: Walter White toma una decisión radical tras ser diagnosticado
              con cáncer de pulmón.
          - numero: 1
            titulo: Piloto
            duracion: 58
            id: 1
            sinopsis: Walter White toma una decisión radical tras ser diagnosticado
              con cáncer de pulmón.
        id: 1
        nombre: Juan Miguel León Rojas
        pais: España
      properties:
        id:
          example: 1
          format: int64
          title: id
          type: integer
        nombre:
          example: Juan Miguel León Rojas
          title: nombre
          type: string
        correo_electronico:
          example: juan@gmail.com
          format: email
          title: correo_electronico
          type: string
        pais:
          example: España
          title: pais
          type: string
        plan_suscripcion:
          enum:
          - Basico
          - Estandar
          - Premium
          example: Básico
          title: plan_suscripcion
          type: string
        dispositivos:
          example:
          - Móvil
          - TV
          items:
            type: string
          title: dispositivos
          type: array
        perfiles:
          items:
            $ref: '#/components/schemas/Perfil'
          title: perfiles
          type: array
      required:
      - correo_electronico
      - nombre
      - plan_suscripcion
      title: Usuario
      type: object
    UsuarioUpdate:
      allOf:
      - $ref: '#/components/schemas/Usuario'
      - required:
        - id
      example:
        dispositivos:
        - Móvil
        - TV
        correo_electronico: juan@gmail.com
        plan_suscripcion: Básico
        perfiles:
        - user_id: 3
          preferencias_contenido:
            languages:
            - Espanol
            genres:
            - Action
            - Supernatural
          foto_perfil: path/to/image
          id: 1
          nombre: Euler
          historial_vistos:
          - numero: 1
            titulo: Piloto
            duracion: 58
            id: 1
            sinopsis: Walter White toma una decisión radical tras ser diagnosticado
              con cáncer de pulmón.
          - numero: 1
            titulo: Piloto
            duracion: 58
            id: 1
            sinopsis: Walter White toma una decisión radical tras ser diagnosticado
              con cáncer de pulmón.
          mi_lista:
          - numero: 1
            titulo: Piloto
            duracion: 58
            id: 1
            sinopsis: Walter White toma una decisión radical tras ser diagnosticado
              con cáncer de pulmón.
          - numero: 1
            titulo: Piloto
            duracion: 58
            id: 1
            sinopsis: Walter White toma una decisión radical tras ser diagnosticado
              con cáncer de pulmón.
        - user_id: 3
          preferencias_contenido:
            languages:
            - Espanol
            genres:
            - Action
            - Supernatural
          foto_perfil: path/to/image
          id: 1
          nombre: Euler
          historial_vistos:
          - numero: 1
            titulo: Piloto
            duracion: 58
            id: 1
            sinopsis: Walter White toma una decisión radical tras ser diagnosticado
              con cáncer de pulmón.
          - numero: 1
            titulo: Piloto
            duracion: 58
            id: 1
            sinopsis: Walter White toma una decisión radical tras ser diagnosticado
              con cáncer de pulmón.
          mi_lista:
          - numero: 1
            titulo: Piloto
            duracion: 58
            id: 1
            sinopsis: Walter White toma una decisión radical tras ser diagnosticado
              con cáncer de pulmón.
          - numero: 1
            titulo: Piloto
            duracion: 58
            id: 1
            sinopsis: Walter White toma una decisión radical tras ser diagnosticado
              con cáncer de pulmón.
        id: 1
        nombre: Juan Miguel León Rojas
        pais: España
      title: UsuarioUpdate
    Perfil:
      example:
        user_id: 3
        preferencias_contenido:
          languages:
          - Espanol
          genres:
          - Action
          - Supernatural
        foto_perfil: path/to/image
        id: 1
        nombre: Euler
        historial_vistos:
        - numero: 1
          titulo: Piloto
          duracion: 58
          id: 1
          sinopsis: Walter White toma una decisión radical tras ser diagnosticado
            con cáncer de pulmón.
        - numero: 1
          titulo: Piloto
          duracion: 58
          id: 1
          sinopsis: Walter White toma una decisión radical tras ser diagnosticado
            con cáncer de pulmón.
        mi_lista:
        - numero: 1
          titulo: Piloto
          duracion: 58
          id: 1
          sinopsis: Walter White toma una decisión radical tras ser diagnosticado
            con cáncer de pulmón.
        - numero: 1
          titulo: Piloto
          duracion: 58
          id: 1
          sinopsis: Walter White toma una decisión radical tras ser diagnosticado
            con cáncer de pulmón.
      properties:
        id:
          description: ID del perfil en específico
          example: 1
          format: int64
          title: id
          type: integer
        user_id:
          description: ID del usuario al que pertenece el perfil
          example: 3
          format: int64
          title: user_id
          type: integer
        nombre:
          example: Euler
          title: nombre
          type: string
        foto_perfil:
          example: path/to/image
          title: foto_perfil
          type: string
        historial_vistos:
          items:
            $ref: '#/components/schemas/Capitulo'
          title: historial_vistos
          type: array
        mi_lista:
          description: Lista con el contenido que el usuario ha guardado
          items:
            $ref: '#/components/schemas/Capitulo'
          title: mi_lista
          type: array
        preferencias_contenido:
          $ref: '#/components/schemas/Perfil_preferencias_contenido'
      required:
      - nombre
      title: Perfil
      type: object
...
    Capitulo:
      example:
        numero: 1
        titulo: Piloto
        duracion: 58
        id: 1
        sinopsis: Walter White toma una decisión radical tras ser diagnosticado con
          cáncer de pulmón.
      properties:
        id:
          example: 1
          format: int64
          title: id
          type: integer
        numero:
          example: 1
          format: int32
          title: numero
          type: integer
        titulo:
          example: Piloto
          title: titulo
          type: string
        duracion:
          description: Duración del capítulo en minutos
          example: 58
          format: int32
          title: duracion
          type: integer
        sinopsis:
          example: Walter White toma una decisión radical tras ser diagnosticado con
            cáncer de pulmón.
          title: sinopsis
          type: string
      required:
      - duracion
      - numero
      - titulo
      title: Capitulo
      type: object
...
    Perfil_preferencias_contenido:
      description: Preferencias de contenido del usuario
      example:
        languages:
        - Espanol
        genres:
        - Action
        - Supernatural
      properties:
        genres:
          example:
          - Action
          - Supernatural
          items:
            type: string
          title: genres
          type: array
        languages:
          example:
          - Espanol
          items:
            enum:
            - Ingles
            - Espanol
            type: string
          title: languages
          type: array
      title: Perfil_preferencias_contenido
      type: object